Note on including custom msg pkg in host ws (#192)
It's worth stating that the custom types package needs to exist in the mcu firmware directory as well as the host ws. Using version control is obviously recommended here. Is there a way to better manage this locally without version control and without manually copying changes between the two workspaces?